+++ src/main/kotlin/pullrequestfactory/io/programs/Programs.kt
@@ -23,16 +23,12 @@ object Programs {
             pa.has_version_option() -> ShowVersionOutputProgram(appProps)
             pa.has_open_command_help_option() -> ShowOpenCommandHelpOutputProgram()
             pa.has_invalid_open_command() -> ShowInvalidOpenCommandOutputProgram()
-            pa.has_open_command_in_interactive_mode() -> OpenPullRequestsProgramInteractiveMode(ui, baseUrl, repoUrl, authTokenFromProps)
-            pa.has_open_command() -> {
-                OpenPullRequestsProgram(ui, pa, repoUrl, g.github_api_client(), g.http_client_stats())
-            }
+            pa.has_open_command_in_interactive_mode() -> OpenPullRequestsProgramsInteractiveMode(ui, pa, baseUrl, repoUrl, authTokenFromProps)
+            pa.has_open_command() -> OpenPullRequestsProgram(ui, pa, repoUrl, g.github_api_client(), g.http_client_stats())
             pa.has_close_command_help_option() -> ShowCloseCommandHelpOutputProgram()
             pa.has_invalid_close_command() -> ShowInvalidCloseCommandOutputProgram()
             pa.has_close_command_in_interactive_mode() -> ClosePullRequestsProgramInteractiveMode(ui, baseUrl, repoUrl, authTokenFromProps)
-            pa.has_close_command() -> {
-                ClosePullRequestsPrograms(ui, pa, repoUrl, g.github_api_client(), g.http_client_stats())
-            }
+            pa.has_close_command() -> ClosePullRequestsPrograms(ui, pa, repoUrl, g.github_api_client(), g.http_client_stats())
             else -> ShowHelpOutputProgram()
         }
     }
@@ -40,7 +36,7 @@ object Programs {
 }
 
 class GitHubHttp(
-        private val programArgs: ProgramArgs,
+        programArgs: ProgramArgs,
         private val baseUrl: String
 ) {