+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adhere to https://github.com/olivierlacan/keep-a-changelog.
 - [Issue-#18] Use GitHub's OAuth 2.0 token for Authorization
 - [Issue-#22] Make `OPEN` commands' `-g` option optional when GitHub token is set in user.properties
 - [Issue-#25] Make `CLOSE` commands' `-g` option optional when GitHub token is set in user.properties
+- [Issue-#27] Support `-l` (last iteration is finished) option for interactive `OPEN` command
 
 ### Removed
 
@@ -60,3 +61,4 @@ adhere to https://github.com/olivierlacan/keep-a-changelog.
 [Issue-#20]: https://github.com/ClausPolanka/github-pr-factory/issues/20
 [Issue-#22]: https://github.com/ClausPolanka/github-pr-factory/issues/22
 [Issue-#25]: https://github.com/ClausPolanka/github-pr-factory/issues/25
+[Issue-#27]: https://github.com/ClausPolanka/github-pr-factory/issues/27